Extracting `randomWait()` method
As I was looking through the code, I saw several instances of randomWait
definitions.  This change consolidates those into a single utility
function.  It also narrows

Along the way, I discovered that I had broken the `make newdb` task.
Hence the change for the `getIngestFilePathsByYear` path.
