+++ ingest/loadFunders.ts
@@ -13,6 +13,7 @@ import insertSubfunderNameVariances from './gql/insertSubfunderNameVariances'
 import { __EnumValue } from 'graphql'
 import dotenv from 'dotenv'
 import pMap from 'p-map'
+import { randomWait } from './units/randomWait'
 
 dotenv.config({
   path: '../.env'
@@ -33,18 +34,6 @@ const client = new ApolloClient({
   }),
   cache: new InMemoryCache()
 })
-
-async function wait(ms){
-  return new Promise((resolve, reject)=> {
-    setTimeout(() => resolve(true), ms );
-  });
-}
-
-async function randomWait(seedTime, index){
-  const waitTime = 1000 * (index % 5)
-  //console.log(`Thread Waiting for ${waitTime} ms`)
-  await wait(waitTime)
-}
 
 function createSubfunderNameVarianceObjects (newSubfundersFromCSV, existingSubfundersByShortName) {
   let newFunderNamevariances = []
